added unsubscribe user for course test added unsubscribe user for course validation refacoring fixed several bugs

+++ dokeos/application/lib/weblcms/webservices/webservices_course.class.php
@@ -212,7 +212,7 @@ class WebServicesCourse
         if($this->webservice->can_execute($input_course, 'subscribe user'))
 		{
             unset($input_course[hash]);            
-            if($this->validator->validate_subscribe_or_unsubscribe_user($input_course)) //input validation
+            if($this->validator->validate_subscribe_user($input_course)) //input validation
             {                               
                 $cur = new CourseUserRelation($input_course[course_code],$input_course[user_id]);                
                 return $this->webservice->raise_message($cur->create());
@@ -233,12 +233,17 @@ class WebServicesCourse
 	{
 		if($this->webservice->can_execute($input_course, 'unsubscribe user'))
 		{
-            unset($input_course[hash]);
-            $cur = new CourseUserRelation($input_course[visual_code],$input_course[user_id]);
-            unset($input_course['user_id']);
-            unset($input_course['visual_code']);
-            $cur->set_default_properties($input_course);
-            return $this->webservice->raise_message($cur->delete());
+           unset($input_course[hash]);
+           if($this->validator->validate_unsubscribe_user($input_course)) //input validation
+            {
+                $cur = new CourseUserRelation($input_course[course_code],$input_course[user_id]);
+                return $this->webservice->raise_message($cur->delete());
+            }
+            else
+            {
+                return $this->webservice->raise_error('Could not update course. Please check the data you\'ve provided.');
+            }
+           
         }
         else
         {
